workshop 3 - granual control over endpoints

+++ WebDev/5-NodeJS-Express-MongoDB/nucampsiteServer/routes/campsiteRouter.js
@@ -15,10 +15,10 @@ campsiteRouter
             })
             .catch((err) => next(err));
     })
-    .post(authenticate.verifyUser, (req, res, next) => {
+    .post(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
         Campsite.create(req.body)
             .then((campsite) => {
-                console.log("Campsite Created ", campsite);
+                //console.log("Campsite Created ", campsite);
                 res.statusCode = 200;
                 res.setHeader("Content-Type", "application/json");
                 res.json(campsite);
@@ -29,7 +29,7 @@ campsiteRouter
         res.statusCode = 403;
         res.end("PUT operation not supported on /campsites");
     })
-    .delete(authenticate.verifyUser, (req, res, next) => {
+    .delete(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
         Campsite.deleteMany()
             .then((response) => {
                 res.statusCode = 200;
@@ -50,11 +50,11 @@ campsiteRouter
             })
             .catch((err) => next(err));
     })
-    .post((req, res) => {
+    .post(authenticate.verifyUser, (req, res) => {
         res.statusCode = 403;
         res.end(`POST operation not supported on /campsites/${req.params.campsiteId}`);
     })
-    .put((req, res, next) => {
+    .put(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
         Campsite.findByIdAndUpdate(req.params.campsiteId, { $set: req.body }, { new: true })
             .then((campsite) => {
                 res.statusCode = 200;
@@ -63,7 +63,7 @@ campsiteRouter
             })
             .catch((err) => next(err));
     })
-    .delete((req, res, next) => {
+    .delete(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
         Campsite.findByIdAndDelete(req.params.campsiteId)
             .then((response) => {
                 res.statusCode = 200;
@@ -83,7 +83,7 @@ campsiteRouter
                     res.setHeader("Content-Type", "application/json");
                     res.json(campsite.comments);
                 } else {
-                    err = new Error(`Campsite ${req.params.campsiteId} not found`);
+                    const err = new Error(`Campsite ${req.params.campsiteId} not found`);
                     err.status = 404;
                     return next(err);
                 }
@@ -94,6 +94,7 @@ campsiteRouter
         Campsite.findById(req.params.campsiteId)
             .then((campsite) => {
                 if (campsite) {
+                    req.body.author = req.user._id;
                     campsite.comments.push(req.body);
                     campsite
                         .save()
@@ -104,7 +105,7 @@ campsiteRouter
                         })
                         .catch((err) => next(err));
                 } else {
-                    err = new Error(`Campsite ${req.params.campsiteId} not found`);
+                    const err = new Error(`Campsite ${req.params.campsiteId} not found`);
                     err.status = 404;
                     return next(err);
                 }
@@ -115,7 +116,7 @@ campsiteRouter
         res.statusCode = 403;
         res.end(`PUT operation not supported on /campsites/${req.params.campsiteId}/comments`);
     })
-    .delete(authenticate.verifyUser, (req, res, next) => {
+    .delete(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
         Campsite.findById(req.params.campsiteId)
             .then((campsite) => {
                 if (campsite) {
@@ -131,7 +132,7 @@ campsiteRouter
                         })
                         .catch((err) => next(err));
                 } else {
-                    err = new Error(`Campsite ${req.params.campsiteId} not found`);
+                    const err = new Error(`Campsite ${req.params.campsiteId} not found`);
                     err.status = 404;
                     return next(err);
                 }
@@ -149,11 +150,11 @@ campsiteRouter
                     res.setHeader("Content-Type", "application/json");
                     res.json(campsite.comments.id(req.params.commentId));
                 } else if (!campsite) {
-                    err = new Error(`Campsite ${req.params.campsiteId} not found`);
+                    const err = new Error(`Campsite ${req.params.campsiteId} not found`);
                     err.status = 404;
                     return next(err);
                 } else {
-                    err = new Error(`Comment ${req.params.commentId} not found`);
+                    const err = new Error(`Comment ${req.params.commentId} not found`);
                     err.status = 404;
                     return next(err);
                 }
@@ -170,26 +171,35 @@ campsiteRouter
         Campsite.findById(req.params.campsiteId)
             .then((campsite) => {
                 if (campsite && campsite.comments.id(req.params.commentId)) {
-                    if (req.body.rating) {
-                        campsite.comments.id(req.params.commentId).rating = req.body.rating;
+                    const authorId = campsite.comments.id(req.params.commentId).author;
+                    const userId = req.user._id;
+
+                    if (authorId.toString() === userId.toString()) {
+                        if (req.body.rating) {
+                            campsite.comments.id(req.params.commentId).rating = req.body.rating;
+                        }
+                        if (req.body.text) {
+                            campsite.comments.id(req.params.commentId).text = req.body.text;
+                        }
+                        campsite
+                            .save()
+                            .then((campsite) => {
+                                res.statusCode = 200;
+                                res.setHeader("Content-Type", "application/json");
+                                res.json(campsite);
+                            })
+                            .catch((err) => next(err));
+                    } else {
+                        const err = new Error(`Not your comment to edit`);
+                        err.status = 403;
+                        return next(err);
                     }
-                    if (req.body.text) {
-                        campsite.comments.id(req.params.commentId).text = req.body.text;
-                    }
-                    campsite
-                        .save()
-                        .then((campsite) => {
-                            res.statusCode = 200;
-                            res.setHeader("Content-Type", "application/json");
-                            res.json(campsite);
-                        })
-                        .catch((err) => next(err));
                 } else if (!campsite) {
-                    err = new Error(`Campsite ${req.params.campsiteId} not found`);
+                    const err = new Error(`Campsite ${req.params.campsiteId} not found`);
                     err.status = 404;
                     return next(err);
                 } else {
-                    err = new Error(`Comment ${req.params.commentId} not found`);
+                    const err = new Error(`Comment ${req.params.commentId} not found`);
                     err.status = 404;
                     return next(err);
                 }
@@ -200,21 +210,30 @@ campsiteRouter
         Campsite.findById(req.params.campsiteId)
             .then((campsite) => {
                 if (campsite && campsite.comments.id(req.params.commentId)) {
-                    campsite.comments.id(req.params.commentId).deleteOne();
-                    campsite
-                        .save()
-                        .then((campsite) => {
-                            res.statusCode = 200;
-                            res.setHeader("Content-Type", "application/json");
-                            res.json(campsite);
-                        })
-                        .catch((err) => next(err));
+                    const authorId = campsite.comments.id(req.params.commentId).author;
+                    const userId = req.user._id;
+
+                    if (authorId.toString() === userId.toString()) {
+                        campsite.comments.id(req.params.commentId).deleteOne();
+                        campsite
+                            .save()
+                            .then((campsite) => {
+                                res.statusCode = 200;
+                                res.setHeader("Content-Type", "application/json");
+                                res.json(campsite);
+                            })
+                            .catch((err) => next(err));
+                    } else {
+                        const err = new Error(`Not your comment to delete`);
+                        err.status = 403;
+                        return next(err);
+                    }
                 } else if (!campsite) {
-                    err = new Error(`Campsite ${req.params.campsiteId} not found`);
+                    const err = new Error(`Campsite ${req.params.campsiteId} not found`);
                     err.status = 404;
                     return next(err);
                 } else {
-                    err = new Error(`Comment ${req.params.commentId} not found`);
+                    const err = new Error(`Comment ${req.params.commentId} not found`);
                     err.status = 404;
                     return next(err);
                 }
